requirements.txt
# Ron Clanker FPL Management System - Dependencies

# Core
python>=3.11

# Database
sqlite3  # Built-in with Python
psycopg2-binary>=2.9.9  # PostgreSQL adapter
redis>=5.0.0  # Redis client for event bus

# Event-driven architecture
celery>=5.3.4  # Task queue and scheduler

# Async support
asyncio  # Built-in with Python

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# HTTP requests
requests>=2.31.0
aiohttp>=3.9.0

# Configuration
python-dotenv>=1.0.0  # Load .env files

# Telegram bot
python-telegram-bot>=13.15

# Intelligence gathering
feedparser>=6.0.10  # RSS feed parsing
python-dateutil>=2.8.2  # Date parsing for RSS
beautifulsoup4>=4.12.0  # HTML parsing for web scraping
youtube-transcript-api>=0.6.1  # YouTube transcript extraction
rapidfuzz>=3.5.0  # Fuzzy string matching for player names

# FPL Data (MCP Server)
fpl-mcp>=0.1.0

# Logging
python-json-logger>=2.0.7

# ML/AI
# torch>=2.0.0  # Future: Deep learning
scikit-learn>=1.3.0  # Machine learning models
xgboost>=2.0.0  # Gradient boosting (ensemble stacking)
joblib>=1.3.0  # Model serialization

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code quality
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

# Development
ipython>=8.14.0
jupyter>=1.0.0
anthropic>=0.40.0
